+++ src/kernel/idt.h
@@ -2,6 +2,6 @@
 
 #include <stdint.h>
 
-void set_idt_gate(uint32_t n, uint32_t handler);
+void set_idt_gate(uint32_t n, uint64_t handler);
 
 void set_idt(void);
